commit before changing code to accomadate buffered calling of work function(buffers 4096 samples))
Testing the work function reveals that the a buffer of 4096 is gathered before the function is called. The code is made only to take 1 value from the 4096, this means that we have been skipping 4095 values. This also explains why the result takes so much time to output. The plan is to shift the code to a circular queue in order to accomadate taking every sample. Then the change in performance can be analysed.