Code works with Noise too
At this point, the code is able to predict the delay even with good amount of noise.
However, a muxxed stream of changing delays(2 vector lengths) does not work. Have to see why.
Primary questions at this point:
Why did the code work before when we took only a single value from the input buffer, though it is now noticeably faster.

What is accuracy of the code over different delay values and different noise values and different samples rates